+++ mediation/InMobi/CHANGELOG.md
@@ -1,5 +1,16 @@
 # InMobi Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 2.1.0
+
+- Supports Android adapter version 7.1.0.0.
+- Supports iOS adapter version 7.1.1.1.
+- Added the `InMobi.UpdateGDPRConsent()` method.
+
+## 2.0.0
+
+- Supports Android adapter version 7.0.4.0.
+- Supports iOS adapter version 7.1.1.0.
+
 ## 1.0.0
 
 - First release!